fix long just

+++ Lesson6/Lesson6/GameViewController.m
@@ -63,8 +63,8 @@
                               forState:UIControlStateNormal];
         cardButton.enabled = !card.isMatched;
         self.cardLogLabel.text = [NSString stringWithString:self.game.matchLog];
-        self.scoreLabel.text = [NSString stringWithFormat:@"Total score: %li", self.game.score];
-        self.gameLogLabel.text = [NSString stringWithFormat:@"Opened %@, current score: %li", self.currentPlayCard.contents, self.game.currentScore];
+        self.scoreLabel.text = [NSString stringWithFormat:@"Total score: %li", (long)self.game.score];
+        self.gameLogLabel.text = [NSString stringWithFormat:@"Opened %@, current score: %li", self.currentPlayCard.contents, (long)self.game.currentScore];
     }
     
     if (self.game.gameFinished) {
@@ -102,7 +102,7 @@
     
     self.cardLogLabel.lineBreakMode = NSLineBreakByWordWrapping;
     self.cardLogLabel.numberOfLines = 0;
-    self.scoreLabel.text = [NSString stringWithFormat:@"Total score: %li", self.game.score];
+    self.scoreLabel.text = [NSString stringWithFormat:@"Total score: %li", (long)self.game.score];
     self.gameLogLabel.text = @"Log is empty.";
     self.cardLogLabel.text = @"Card log is empty.";
